updated to refactor Ditto Manager and update observers
