+++ SwiftUI/Edge Studio/Data/SelectedApp/DittoManager_Import.swift
@@ -1,0 +1,13 @@
+//
+//  DittoManager_Import.swift
+//  Edge Studio
+//
+//  Created by Aaron LaBeau on 6/4/25.
+//
+
+import Foundation
+
+// MARK: Ditto Selected App - Import Operations
+extension DittoManager {
+    
+}
