+++ SwiftUI/Edge Studio/Views/Tabs/ObservablesTabView.swift
@@ -90,7 +90,28 @@ struct ObservablesTabView: View {
         .sheet(
            isPresented: $viewModel.isEditorPresented,
         ){
-            
+            ObserverEditorView(
+                isPresented: $viewModel.isEditorPresented,
+                selectedObservable: viewModel.selectedObservable ?? DittoObservable.new()
+            )
+            #if os(macOS)
+            .frame(minWidth: 860,
+                   idealWidth: 1000,
+                   maxWidth: 1080,
+                   minHeight: 500,
+                   idealHeight: 800)
+            #elseif os(iOS)
+            .frame(
+                minWidth: UIDevice.current.userInterfaceIdiom == .pad ? 600 : nil,
+                idealWidth: UIDevice.current.userInterfaceIdiom == .pad ? 1000 : nil,
+                maxWidth: UIDevice.current.userInterfaceIdiom == .pad ? 1080 : nil,
+                minHeight: UIDevice.current.userInterfaceIdiom == .pad ? 500 : nil,
+                idealHeight: UIDevice.current.userInterfaceIdiom == .pad ? 700 : nil,
+                maxHeight: UIDevice.current.userInterfaceIdiom == .pad ? 800 : nil
+            )
+            #endif
+            .environmentObject(appState)
+            .presentationDetents([.medium, .large])
         }
         .toolbar {
             #if os(iPadOS)
@@ -99,7 +120,7 @@ struct ObservablesTabView: View {
                 }
                 ToolbarItem(placement: .topBarTrailing) {
                     Button {
-                        viewModel.selectedObservable = DittoObservable.new()
+                        viewModel.showObserverEditor(DittoObservable.new())
                     } label: {
                         Image(systemName: "plus")
                     }
@@ -109,7 +130,7 @@ struct ObservablesTabView: View {
             #if os(macOS)
             ToolbarItem(id: "add", placement: .primaryAction) {
                 Button {
-                    viewModel.selectedObservable = DittoObservable.new()
+                    viewModel.showObserverEditor(DittoObservable.new())
                 } label: {
                     Image(systemName: "plus")
                 }
@@ -145,6 +166,11 @@ extension ObservablesTabView {
         
         init(_ dittoAppConfig: DittoAppConfig) {
             self.selectedApp = dittoAppConfig
+        }
+        
+        func showObserverEditor(_ observable: DittoObservable) {
+            selectedObservable = observable
+            isEditorPresented = true
         }
         
         func closeSelectedApp() async {
